feat: move analytics tab to analytics view
