+++ src/components/Analytics/AnalyticsView.tsx
@@ -1,0 +1,11 @@
+import { Box } from '@mui/material';
+
+import AnalyticsMenu from './AnalyticsMenu';
+
+const AnalyticsView = (): JSX.Element => (
+  <Box sx={{ width: '100%' }}>
+    <AnalyticsMenu />
+  </Box>
+);
+
+export default AnalyticsView;
