+++ src/components/navigation/AdminView.tsx
@@ -5,12 +5,10 @@ import { Tab, Tabs } from '@mui/material';
 import Box from '@mui/material/Box';
 
 import {
-  NAVIGATION_ANALYTICS_BUTTON_CY,
   NAVIGATION_CREATE_QUIZ_BUTTON_CY,
   NAVIGATION_RESULT_BUTTON_CY,
   NAVIGATION_TAB_CONTAINER_CY,
 } from '../../config/selectors';
-import AnalyticsMenu from '../Analytics/AnalyticsMenu';
 import CreateView from '../create/CreateView';
 import ResultTables from '../results/ResultTables';
 import TabPanel from './TabPanel';
@@ -37,10 +35,6 @@ const AdminView = () => {
             data-cy={NAVIGATION_CREATE_QUIZ_BUTTON_CY}
           />
           <Tab label={t('Results')} data-cy={NAVIGATION_RESULT_BUTTON_CY} />
-          <Tab
-            label={t('Analytics')}
-            data-cy={NAVIGATION_ANALYTICS_BUTTON_CY}
-          />
         </Tabs>
       </Box>
       <TabPanel tab={tab} index={0}>
@@ -48,9 +42,6 @@ const AdminView = () => {
       </TabPanel>
       <TabPanel tab={tab} index={1}>
         <ResultTables headerElem={headerElem} />
-      </TabPanel>
-      <TabPanel tab={tab} index={2}>
-        <AnalyticsMenu headerElem={headerElem} />
       </TabPanel>
     </Box>
   );
